fixed java team folder
